# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/parts": {
      "post": {
        "operationId": "PartsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Part created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new part",
        "tags": [
          "Parts"
        ]
      },
      "get": {
        "operationId": "PartsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of parts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Retrieve all parts",
        "tags": [
          "Parts"
        ]
      }
    },
    "/parts/{id}": {
      "get": {
        "operationId": "PartsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Part found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Retrieve a part by its UUID",
        "tags": [
          "Parts"
        ]
      },
      "put": {
        "operationId": "PartsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Part updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a part by its UUID",
        "tags": [
          "Parts"
        ]
      },
      "delete": {
        "operationId": "PartsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for deletion"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Part deleted"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a part by its UUID, with reason",
        "tags": [
          "Parts"
        ]
      }
    },
    "/parts/{id}/claim": {
      "post": {
        "operationId": "PartsController_claimPart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Part claimed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Claim a part for the current user",
        "tags": [
          "Parts",
          "Parts",
          "Claims"
        ]
      }
    },
    "/parts/{id}/request": {
      "post": {
        "operationId": "PartsController_requestPart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Part requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Request a part for the current user",
        "tags": [
          "Parts",
          "Parts",
          "Requests"
        ]
      }
    },
    "/parts/{id}/return": {
      "post": {
        "operationId": "PartsController_returnPart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Reason for return"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Part returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Return a claimed part for the current user",
        "tags": [
          "Parts",
          "Parts",
          "Returns"
        ]
      }
    },
    "/parts/{id}/collect": {
      "post": {
        "operationId": "PartsController_collectPart",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Part UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Part collected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Part not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mark a claimed part as collected",
        "tags": [
          "Parts",
          "Parts",
          "Collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Parts Management API",
    "description": "Endpoints to manage parts, claims, and users",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePartDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "partNumber": {
            "type": "string"
          },
          "forDeviceModels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PENDING_COLLECTION",
              "CLAIMED",
              "COLLECTED",
              "UNAVAILABLE"
            ]
          }
        },
        "required": [
          "name",
          "partNumber",
          "forDeviceModels",
          "quantity"
        ]
      },
      "PartResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "partNumber": {
            "type": "string"
          },
          "forDeviceModels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quantity": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PENDING_COLLECTION",
              "CLAIMED",
              "COLLECTED",
              "UNAVAILABLE"
            ]
          },
          "claimedByName": {
            "type": "object"
          },
          "claimedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAtTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtTimestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "partNumber",
          "forDeviceModels",
          "quantity",
          "status",
          "createdAtTimestamp",
          "updatedAtTimestamp"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code of the error"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short description of the error type"
          },
          "message": {
            "example": [
              "username must be a string",
              "password should not be empty"
            ],
            "description": "Detailed error message(s); either a string or an array of strings",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "example": "2025-09-08T09:30:00.000Z",
            "description": "Timestamp of when the error occurred"
          },
          "path": {
            "type": "string",
            "example": "/auth/login",
            "description": "Request path that triggered the error"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
      "UpdatePartDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "partNumber": {
            "type": "string"
          },
          "forDeviceModels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quantity": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PENDING_COLLECTION",
              "CLAIMED",
              "COLLECTED",
              "UNAVAILABLE"
            ]
          }
        }
      },
...
    }
  }
}